Refactored: Extracted Ortb2RequestFactory
Refactored workflow to have more granular control of each step (to include hooks in between).

!! Calculating account as first step and don't change it after stored request was merged.
So storedRequest can't change account from now on. !!

Ortb2RequestFactory include common logic for Amp/Video/Auction request factory.
AuctionImplicitParametersInjector - needs to be extracted into separate class because it was 80% of Ortb2RequestFactory.
We can extract similar classes for AMP and Video later

Refactored tests.
