+++ src/main/java/org/prebid/server/auction/model/AuctionContext.java
@@ -40,4 +40,15 @@ public class AuctionContext {
     PrivacyContext privacyContext;
 
     GeoInfo geoInfo;
+
+    public AuctionContext with(BidRequest bidRequest) {
+        return this.toBuilder().bidRequest(bidRequest).build();
+    }
+
+    public AuctionContext with(PrivacyContext privacyContext) {
+        return this.toBuilder()
+                .privacyContext(privacyContext)
+                .geoInfo(privacyContext.getTcfContext().getGeoInfo())
+                .build();
+    }
 }
