+++ src/main/java/org/prebid/server/auction/requestfactory/AuctionImplicitParametersInjector.java
@@ -1,19 +1,13 @@
-package org.prebid.server.auction;
+package org.prebid.server.auction.requestfactory;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.iab.openrtb.request.App;
 import com.iab.openrtb.request.BidRequest;
 import com.iab.openrtb.request.Device;
-import com.iab.openrtb.request.Geo;
 import com.iab.openrtb.request.Imp;
-import com.iab.openrtb.request.Publisher;
 import com.iab.openrtb.request.Site;
 import com.iab.openrtb.request.Source;
-import io.netty.buffer.ByteBufInputStream;
-import io.vertx.core.Future;
-import io.vertx.core.buffer.Buffer;
 import io.vertx.core.http.HttpServerRequest;
 import io.vertx.core.logging.Logger;
 import io.vertx.core.logging.LoggerFactory;
@@ -21,27 +15,19 @@ import io.vertx.ext.web.RoutingContext;
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.lang3.ObjectUtils;
 import org.apache.commons.lang3.StringUtils;
-import org.prebid.server.auction.model.AuctionContext;
+import org.prebid.server.auction.ImplicitParametersExtractor;
+import org.prebid.server.auction.IpAddressHelper;
+import org.prebid.server.auction.PriceGranularity;
+import org.prebid.server.auction.TimeoutResolver;
 import org.prebid.server.auction.model.IpAddress;
-import org.prebid.server.cookie.UidsCookieService;
-import org.prebid.server.exception.BlacklistedAccountException;
 import org.prebid.server.exception.BlacklistedAppException;
 import org.prebid.server.exception.InvalidRequestException;
 import org.prebid.server.exception.PreBidException;
-import org.prebid.server.exception.UnauthorizedAccountException;
-import org.prebid.server.execution.Timeout;
-import org.prebid.server.execution.TimeoutFactory;
-import org.prebid.server.geolocation.model.GeoInfo;
 import org.prebid.server.identity.IdGenerator;
 import org.prebid.server.json.JacksonMapper;
-import org.prebid.server.log.ConditionalLogger;
-import org.prebid.server.metric.MetricName;
-import org.prebid.server.privacy.model.PrivacyContext;
 import org.prebid.server.proto.openrtb.ext.request.ExtDevice;
 import org.prebid.server.proto.openrtb.ext.request.ExtMediaTypePriceGranularity;
 import org.prebid.server.proto.openrtb.ext.request.ExtPriceGranularity;
-import org.prebid.server.proto.openrtb.ext.request.ExtPublisher;
-import org.prebid.server.proto.openrtb.ext.request.ExtPublisherPrebid;
 import org.prebid.server.proto.openrtb.ext.request.ExtRequest;
 import org.prebid.server.proto.openrtb.ext.request.ExtRequestPrebid;
 import org.prebid.server.proto.openrtb.ext.request.ExtRequestPrebidCache;
@@ -49,16 +35,9 @@ import org.prebid.server.proto.openrtb.ext.request.ExtRequestPrebidChannel;
 import org.prebid.server.proto.openrtb.ext.request.ExtRequestTargeting;
 import org.prebid.server.proto.openrtb.ext.request.ExtSite;
 import org.prebid.server.proto.openrtb.ext.response.BidType;
-import org.prebid.server.settings.ApplicationSettings;
-import org.prebid.server.settings.model.Account;
-import org.prebid.server.settings.model.AccountStatus;
 import org.prebid.server.util.HttpUtil;
 import org.prebid.server.util.StreamUtil;
-import org.prebid.server.validation.RequestValidator;
-import org.prebid.server.validation.model.ValidationResult;
 
-import java.io.IOException;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.Currency;
@@ -69,14 +48,9 @@ import java.util.Set;
 import java.util.function.Function;
 import java.util.stream.Collectors;
 
-/**
- * Used in OpenRTB request processing.
- */
-public class AuctionRequestFactory {
+public class AuctionImplicitParametersInjector {
 
-    private static final Logger logger = LoggerFactory.getLogger(AuctionRequestFactory.class);
-    private static final ConditionalLogger EMPTY_ACCOUNT_LOGGER = new ConditionalLogger("empty_account", logger);
-    private static final ConditionalLogger UNKNOWN_ACCOUNT_LOGGER = new ConditionalLogger("unknown_account", logger);
+    private static final Logger logger = LoggerFactory.getLogger(AuctionImplicitParametersInjector.class);
 
     private static final String WEB_CHANNEL = "web";
     private static final String APP_CHANNEL = "app";
@@ -87,64 +61,28 @@ public class AuctionRequestFactory {
     private static final Set<String> IMP_EXT_NON_BIDDER_FIELDS = Collections.unmodifiableSet(new HashSet<>(
             Arrays.asList(PREBID_EXT, "context", "all", "general", "skadn", "data")));
 
-    private final long maxRequestSize;
-    private final boolean enforceValidAccount;
     private final boolean shouldCacheOnlyWinningBids;
     private final String adServerCurrency;
     private final List<String> blacklistedApps;
-    private final List<String> blacklistedAccounts;
-    private final StoredRequestProcessor storedRequestProcessor;
     private final ImplicitParametersExtractor paramsExtractor;
     private final IpAddressHelper ipAddressHelper;
-    private final UidsCookieService uidsCookieService;
-    private final RequestValidator requestValidator;
-    private final InterstitialProcessor interstitialProcessor;
-    private final TimeoutResolver timeoutResolver;
-    private final TimeoutFactory timeoutFactory;
-    private final ApplicationSettings applicationSettings;
     private final IdGenerator sourceIdGenerator;
-    private final PrivacyEnforcementService privacyEnforcementService;
     private final JacksonMapper mapper;
-    private final OrtbTypesResolver ortbTypesResolver;
 
-    public AuctionRequestFactory(long maxRequestSize,
-                                 boolean enforceValidAccount,
-                                 boolean shouldCacheOnlyWinningBids,
-                                 String adServerCurrency,
-                                 List<String> blacklistedApps,
-                                 List<String> blacklistedAccounts,
-                                 StoredRequestProcessor storedRequestProcessor,
-                                 ImplicitParametersExtractor paramsExtractor,
-                                 IpAddressHelper ipAddressHelper,
-                                 UidsCookieService uidsCookieService,
-                                 RequestValidator requestValidator,
-                                 InterstitialProcessor interstitialProcessor,
-                                 OrtbTypesResolver ortbTypesResolver,
-                                 TimeoutResolver timeoutResolver,
-                                 TimeoutFactory timeoutFactory,
-                                 ApplicationSettings applicationSettings,
-                                 IdGenerator sourceIdGenerator,
-                                 PrivacyEnforcementService privacyEnforcementService,
-                                 JacksonMapper mapper) {
+    public AuctionImplicitParametersInjector(boolean shouldCacheOnlyWinningBids,
+                                             String adServerCurrency,
+                                             List<String> blacklistedApps,
+                                             ImplicitParametersExtractor paramsExtractor,
+                                             IpAddressHelper ipAddressHelper,
+                                             IdGenerator sourceIdGenerator,
+                                             JacksonMapper mapper) {
 
-        this.maxRequestSize = maxRequestSize;
-        this.enforceValidAccount = enforceValidAccount;
         this.shouldCacheOnlyWinningBids = shouldCacheOnlyWinningBids;
         this.adServerCurrency = validateCurrency(Objects.requireNonNull(adServerCurrency));
         this.blacklistedApps = Objects.requireNonNull(blacklistedApps);
-        this.blacklistedAccounts = Objects.requireNonNull(blacklistedAccounts);
-        this.storedRequestProcessor = Objects.requireNonNull(storedRequestProcessor);
         this.paramsExtractor = Objects.requireNonNull(paramsExtractor);
         this.ipAddressHelper = Objects.requireNonNull(ipAddressHelper);
-        this.uidsCookieService = Objects.requireNonNull(uidsCookieService);
-        this.requestValidator = Objects.requireNonNull(requestValidator);
-        this.interstitialProcessor = Objects.requireNonNull(interstitialProcessor);
-        this.ortbTypesResolver = Objects.requireNonNull(ortbTypesResolver);
-        this.timeoutResolver = Objects.requireNonNull(timeoutResolver);
-        this.timeoutFactory = Objects.requireNonNull(timeoutFactory);
-        this.applicationSettings = Objects.requireNonNull(applicationSettings);
         this.sourceIdGenerator = Objects.requireNonNull(sourceIdGenerator);
-        this.privacyEnforcementService = Objects.requireNonNull(privacyEnforcementService);
         this.mapper = Objects.requireNonNull(mapper);
     }
 
@@ -161,110 +99,14 @@ public class AuctionRequestFactory {
     }
 
     /**
-     * Creates {@link AuctionContext} based on {@link RoutingContext}.
-     */
-    public Future<AuctionContext> fromRequest(RoutingContext routingContext, long startTime) {
-        final List<String> errors = new ArrayList<>();
-        final BidRequest incomingBidRequest;
-        try {
-            incomingBidRequest = parseRequest(routingContext, errors);
-        } catch (InvalidRequestException e) {
-            return Future.failedFuture(e);
-        }
-
-        return updateBidRequest(routingContext, incomingBidRequest)
-                .compose(bidRequest -> toAuctionContext(
-                        routingContext,
-                        bidRequest,
-                        requestTypeMetric(bidRequest),
-                        errors,
-                        startTime,
-                        timeoutResolver));
-    }
-
-    /**
-     * Returns filled out {@link AuctionContext} based on given arguments.
-     * <p>
-     * Note: {@link TimeoutResolver} used here as argument because this method is utilized in AMP processing.
-     */
-    Future<AuctionContext> toAuctionContext(RoutingContext routingContext,
-                                            BidRequest bidRequest,
-                                            MetricName requestTypeMetric,
-                                            List<String> errors,
-                                            long startTime,
-                                            TimeoutResolver timeoutResolver) {
-
-        final Timeout timeout = timeout(bidRequest, startTime, timeoutResolver);
-
-        return accountFrom(bidRequest, timeout, routingContext)
-                .compose(account -> privacyEnforcementService.contextFromBidRequest(
-                        bidRequest, account, requestTypeMetric, timeout, errors)
-                        .map(privacyContext -> AuctionContext.builder()
-                                .routingContext(routingContext)
-                                .uidsCookie(uidsCookieService.parseFromRequest(routingContext))
-                                .bidRequest(enrichBidRequestWithAccountAndPrivacyData(
-                                        bidRequest, account, privacyContext))
-                                .requestTypeMetric(requestTypeMetric)
-                                .timeout(timeout)
-                                .account(account)
-                                .prebidErrors(errors)
-                                .debugWarnings(new ArrayList<>())
-                                .privacyContext(privacyContext)
-                                .geoInfo(privacyContext.getTcfContext().getGeoInfo())
-                                .build()));
-    }
-
-    /**
-     * Parses request body to {@link BidRequest}.
-     * <p>
-     * Throws {@link InvalidRequestException} if body is empty, exceeds max request size or couldn't be deserialized.
-     */
-    private BidRequest parseRequest(RoutingContext context, List<String> errors) {
-        final Buffer body = context.getBody();
-        if (body == null) {
-            throw new InvalidRequestException("Incoming request has no body");
-        }
-
-        if (body.length() > maxRequestSize) {
-            throw new InvalidRequestException(
-                    String.format("Request size exceeded max size of %d bytes.", maxRequestSize));
-        }
-
-        final JsonNode bidRequestNode;
-        try (ByteBufInputStream inputStream = new ByteBufInputStream(body.getByteBuf())) {
-            bidRequestNode = mapper.mapper().readTree(inputStream);
-        } catch (IOException e) {
-            throw new InvalidRequestException(String.format("Error decoding bidRequest: %s", e.getMessage()));
-        }
-
-        final String referer = paramsExtractor.refererFrom(context.request());
-        ortbTypesResolver.normalizeBidRequest(bidRequestNode, errors, referer);
-
-        try {
-            return mapper.mapper().treeToValue(bidRequestNode, BidRequest.class);
-        } catch (JsonProcessingException e) {
-            throw new InvalidRequestException(String.format("Error decoding bidRequest: %s", e.getMessage()));
-        }
-    }
-
-    /**
-     * Sets {@link BidRequest} properties which were not set explicitly by the client, but can be
-     * updated by values derived from headers and other request attributes.
-     */
-    private Future<BidRequest> updateBidRequest(RoutingContext context, BidRequest bidRequest) {
-        return storedRequestProcessor.processStoredRequests(accountIdFrom(bidRequest), bidRequest)
-                .map(resolvedBidRequest -> fillImplicitParameters(resolvedBidRequest, context, timeoutResolver))
-                .map(this::validateRequest)
-                .map(interstitialProcessor::process);
-    }
-
-    /**
      * If needed creates a new {@link BidRequest} which is a copy of original but with some fields set with values
      * derived from request parameters (headers, cookie etc.).
      * <p>
      * Note: {@link TimeoutResolver} used here as argument because this method is utilized in AMP processing.
      */
-    BidRequest fillImplicitParameters(BidRequest bidRequest, RoutingContext context, TimeoutResolver timeoutResolver) {
+    BidRequest fillImplicitParameters(BidRequest bidRequest,
+                                      RoutingContext context,
+                                      TimeoutResolver timeoutResolver) {
         checkBlacklistedApp(bidRequest);
 
         final BidRequest result;
@@ -854,187 +696,5 @@ public class AuctionRequestFactory {
 
     private static <T, R> R getIfNotNull(T target, Function<T, R> getter) {
         return target != null ? getter.apply(target) : null;
-    }
-
-    /**
-     * Performs thorough validation of fully constructed {@link BidRequest} that is going to be used to hold an auction.
-     */
-    BidRequest validateRequest(BidRequest bidRequest) {
-        final ValidationResult validationResult = requestValidator.validate(bidRequest);
-        if (validationResult.hasErrors()) {
-            throw new InvalidRequestException(validationResult.getErrors());
-        }
-        return bidRequest;
-    }
-
-    /**
-     * Returns {@link Timeout} based on request.tmax and adjustment value of {@link TimeoutResolver}.
-     */
-    private Timeout timeout(BidRequest bidRequest, long startTime, TimeoutResolver timeoutResolver) {
-        final long timeout = timeoutResolver.adjustTimeout(bidRequest.getTmax());
-        return timeoutFactory.create(startTime, timeout);
-    }
-
-    /**
-     * Returns {@link Account} fetched by {@link ApplicationSettings}.
-     */
-    private Future<Account> accountFrom(BidRequest bidRequest, Timeout timeout, RoutingContext routingContext) {
-        final String accountId = accountIdFrom(bidRequest);
-        final boolean blankAccountId = StringUtils.isBlank(accountId);
-
-        if (CollectionUtils.isNotEmpty(blacklistedAccounts) && !blankAccountId
-                && blacklistedAccounts.contains(accountId)) {
-            throw new BlacklistedAccountException(String.format("Prebid-server has blacklisted Account ID: %s, please "
-                    + "reach out to the prebid server host.", accountId));
-        }
-
-        return blankAccountId
-                ? responseForEmptyAccount(routingContext)
-                : applicationSettings.getAccountById(accountId, timeout)
-                .compose(this::ensureAccountActive,
-                        exception -> accountFallback(exception, accountId, routingContext));
-    }
-
-    /**
-     * Extracts publisher id either from {@link BidRequest}.app.publisher or {@link BidRequest}.site.publisher.
-     * If neither is present returns empty string.
-     */
-    private String accountIdFrom(BidRequest bidRequest) {
-        final App app = bidRequest.getApp();
-        final Publisher appPublisher = app != null ? app.getPublisher() : null;
-        final Site site = bidRequest.getSite();
-        final Publisher sitePublisher = site != null ? site.getPublisher() : null;
-
-        final Publisher publisher = ObjectUtils.defaultIfNull(appPublisher, sitePublisher);
-        final String publisherId = publisher != null ? resolvePublisherId(publisher) : null;
-        return ObjectUtils.defaultIfNull(publisherId, StringUtils.EMPTY);
-    }
-
-    /**
-     * Resolves what value should be used as a publisher id - either taken from publisher.ext.parentAccount
-     * or publisher.id in this respective priority.
-     */
-    private String resolvePublisherId(Publisher publisher) {
-        final String parentAccountId = parentAccountIdFromExtPublisher(publisher.getExt());
-        return ObjectUtils.defaultIfNull(parentAccountId, publisher.getId());
-    }
-
-    /**
-     * Parses publisher.ext and returns parentAccount value from it. Returns null if any parsing error occurs.
-     */
-    private String parentAccountIdFromExtPublisher(ExtPublisher extPublisher) {
-        final ExtPublisherPrebid extPublisherPrebid = extPublisher != null ? extPublisher.getPrebid() : null;
-        return extPublisherPrebid != null ? StringUtils.stripToNull(extPublisherPrebid.getParentAccount()) : null;
-    }
-
-    private Future<Account> responseForEmptyAccount(RoutingContext routingContext) {
-        EMPTY_ACCOUNT_LOGGER.warn(accountErrorMessage("Account not specified", routingContext), 100);
-        return responseForUnknownAccount(StringUtils.EMPTY);
-    }
-
-    private static String accountErrorMessage(String message, RoutingContext routingContext) {
-        final HttpServerRequest request = routingContext.request();
-        return String.format("%s, Url: %s and Referer: %s", message, request.absoluteURI(),
-                request.headers().get(HttpUtil.REFERER_HEADER));
-    }
-
-    private Future<Account> accountFallback(Throwable exception, String accountId,
-                                            RoutingContext routingContext) {
-        if (exception instanceof PreBidException) {
-            UNKNOWN_ACCOUNT_LOGGER.warn(accountErrorMessage(exception.getMessage(), routingContext), 100);
-        } else {
-            logger.warn("Error occurred while fetching account: {0}", exception.getMessage());
-            logger.debug("Error occurred while fetching account", exception);
-        }
-
-        // hide all errors occurred while fetching account
-        return responseForUnknownAccount(accountId);
-    }
-
-    private Future<Account> responseForUnknownAccount(String accountId) {
-        return enforceValidAccount
-                ? Future.failedFuture(new UnauthorizedAccountException(
-                String.format("Unauthorized account id: %s", accountId), accountId))
-                : Future.succeededFuture(Account.empty(accountId));
-    }
-
-    private Future<Account> ensureAccountActive(Account account) {
-        final String accountId = account.getId();
-
-        return account.getStatus() == AccountStatus.inactive
-                ? Future.failedFuture(new UnauthorizedAccountException(
-                String.format("Account %s is inactive", accountId), accountId))
-                : Future.succeededFuture(account);
-    }
-
-    private BidRequest enrichBidRequestWithAccountAndPrivacyData(
-            BidRequest bidRequest, Account account, PrivacyContext privacyContext) {
-
-        final ExtRequest requestExt = bidRequest.getExt();
-        final ExtRequest enrichedRequestExt = enrichExtRequest(requestExt, account);
-
-        final Device device = bidRequest.getDevice();
-        final Device enrichedDevice = enrichDevice(device, privacyContext);
-
-        if (enrichedRequestExt != null || enrichedDevice != null) {
-            return bidRequest.toBuilder()
-                    .ext(ObjectUtils.defaultIfNull(enrichedRequestExt, requestExt))
-                    .device(ObjectUtils.defaultIfNull(enrichedDevice, device))
-                    .build();
-        }
-
-        return bidRequest;
-    }
-
-    private ExtRequest enrichExtRequest(ExtRequest ext, Account account) {
-        final ExtRequestPrebid prebidExt = getIfNotNull(ext, ExtRequest::getPrebid);
-        final String integration = getIfNotNull(prebidExt, ExtRequestPrebid::getIntegration);
-        final String accountDefaultIntegration = account.getDefaultIntegration();
-
-        if (StringUtils.isBlank(integration) && StringUtils.isNotBlank(accountDefaultIntegration)) {
-            final ExtRequestPrebid.ExtRequestPrebidBuilder prebidExtBuilder =
-                    prebidExt != null ? prebidExt.toBuilder() : ExtRequestPrebid.builder();
-
-            prebidExtBuilder.integration(accountDefaultIntegration);
-
-            return ExtRequest.of(prebidExtBuilder.build());
-        }
-
-        return null;
-    }
-
-    private Device enrichDevice(Device device, PrivacyContext privacyContext) {
-        final String ipAddress = privacyContext.getIpAddress();
-        final String country = getIfNotNull(privacyContext.getTcfContext().getGeoInfo(), GeoInfo::getCountry);
-
-        final String ipAddressInRequest = getIfNotNull(device, Device::getIp);
-
-        final Geo geo = getIfNotNull(device, Device::getGeo);
-        final String countryFromRequest = getIfNotNull(geo, Geo::getCountry);
-
-        final boolean shouldUpdateIp = ipAddress != null && !Objects.equals(ipAddressInRequest, ipAddress);
-        final boolean shouldUpdateCountry = country != null && !Objects.equals(countryFromRequest, country);
-
-        if (shouldUpdateIp || shouldUpdateCountry) {
-            final Device.DeviceBuilder deviceBuilder = device != null ? device.toBuilder() : Device.builder();
-
-            if (shouldUpdateIp) {
-                deviceBuilder.ip(ipAddress);
-            }
-
-            if (shouldUpdateCountry) {
-                final Geo.GeoBuilder geoBuilder = geo != null ? geo.toBuilder() : Geo.builder();
-                geoBuilder.country(country);
-                deviceBuilder.geo(geoBuilder.build());
-            }
-
-            return deviceBuilder.build();
-        }
-
-        return null;
-    }
-
-    private static MetricName requestTypeMetric(BidRequest bidRequest) {
-        return bidRequest.getApp() != null ? MetricName.openrtb2app : MetricName.openrtb2web;
     }
 }
