+++ src/main/java/org/prebid/server/spring/config/ServiceConfiguration.java
@@ -6,9 +6,7 @@ import io.vertx.core.Vertx;
 import io.vertx.core.file.FileSystem;
 import io.vertx.core.http.HttpClientOptions;
 import io.vertx.core.net.JksOptions;
-import org.prebid.server.auction.AmpRequestFactory;
 import org.prebid.server.auction.AmpResponsePostProcessor;
-import org.prebid.server.auction.AuctionRequestFactory;
 import org.prebid.server.auction.BidResponseCreator;
 import org.prebid.server.auction.BidResponsePostProcessor;
 import org.prebid.server.auction.ExchangeService;
@@ -22,10 +20,14 @@ import org.prebid.server.auction.SchainResolver;
 import org.prebid.server.auction.StoredRequestProcessor;
 import org.prebid.server.auction.StoredResponseProcessor;
 import org.prebid.server.auction.TimeoutResolver;
-import org.prebid.server.auction.VideoRequestFactory;
 import org.prebid.server.auction.VideoResponseFactory;
 import org.prebid.server.auction.VideoStoredRequestProcessor;
 import org.prebid.server.auction.WinningBidComparator;
+import org.prebid.server.auction.requestfactory.AmpRequestFactory;
+import org.prebid.server.auction.requestfactory.AuctionRequestFactory;
+import org.prebid.server.auction.requestfactory.AuctionImplicitParametersInjector;
+import org.prebid.server.auction.requestfactory.Ortb2RequestFactory;
+import org.prebid.server.auction.requestfactory.VideoRequestFactory;
 import org.prebid.server.bidder.BidderCatalog;
 import org.prebid.server.bidder.BidderDeps;
 import org.prebid.server.bidder.BidderErrorNotifier;
@@ -178,48 +180,71 @@ public class ServiceConfiguration {
     }
 
     @Bean
-    AuctionRequestFactory auctionRequestFactory(
-            @Value("${auction.max-request-size}") @Min(0) int maxRequestSize,
-            @Value("${settings.enforce-valid-account}") boolean enforceValidAccount,
+    AuctionImplicitParametersInjector implicitParametersInjector(
             @Value("${auction.cache.only-winning-bids}") boolean shouldCacheOnlyWinningBids,
             @Value("${auction.ad-server-currency}") String adServerCurrency,
             @Value("${auction.blacklisted-apps}") String blacklistedAppsString,
-            @Value("${auction.blacklisted-accounts}") String blacklistedAccountsString,
-            StoredRequestProcessor storedRequestProcessor,
             ImplicitParametersExtractor implicitParametersExtractor,
             IpAddressHelper ipAddressHelper,
-            UidsCookieService uidsCookieService,
-            RequestValidator requestValidator,
-            OrtbTypesResolver ortbTypesResolver,
-            TimeoutResolver timeoutResolver,
-            TimeoutFactory timeoutFactory,
-            ApplicationSettings applicationSettings,
-            PrivacyEnforcementService privacyEnforcementService,
             IdGenerator sourceIdGenerator,
             JacksonMapper mapper) {
 
         final List<String> blacklistedApps = splitToList(blacklistedAppsString);
-        final List<String> blacklistedAccounts = splitToList(blacklistedAccountsString);
 
-        return new AuctionRequestFactory(
-                maxRequestSize,
-                enforceValidAccount,
+        return new AuctionImplicitParametersInjector(
                 shouldCacheOnlyWinningBids,
                 adServerCurrency,
                 blacklistedApps,
-                blacklistedAccounts,
-                storedRequestProcessor,
                 implicitParametersExtractor,
                 ipAddressHelper,
+                sourceIdGenerator,
+                mapper);
+    }
+
+    @Bean
+    Ortb2RequestFactory openRtb2RequestFactory(
+            @Value("${settings.enforce-valid-account}") boolean enforceValidAccount,
+            @Value("${auction.blacklisted-accounts}") String blacklistedAccountsString,
+            UidsCookieService uidsCookieService,
+            RequestValidator requestValidator,
+            TimeoutResolver timeoutResolver,
+            TimeoutFactory timeoutFactory,
+            ApplicationSettings applicationSettings) {
+
+        final List<String> blacklistedAccounts = splitToList(blacklistedAccountsString);
+
+        return new Ortb2RequestFactory(
+                enforceValidAccount,
+                blacklistedAccounts,
                 uidsCookieService,
                 requestValidator,
-                new InterstitialProcessor(),
-                ortbTypesResolver,
                 timeoutResolver,
                 timeoutFactory,
-                applicationSettings,
-                sourceIdGenerator,
+                applicationSettings);
+    }
+
+    @Bean
+    AuctionRequestFactory auctionRequestFactory(
+            @Value("${auction.max-request-size}") @Min(0) int maxRequestSize,
+            Ortb2RequestFactory ortb2RequestFactory,
+            StoredRequestProcessor storedRequestProcessor,
+            ImplicitParametersExtractor implicitParametersExtractor,
+            AuctionImplicitParametersInjector auctionImplicitParametersInjector,
+            OrtbTypesResolver ortbTypesResolver,
+            PrivacyEnforcementService privacyEnforcementService,
+            TimeoutResolver timeoutResolver,
+            JacksonMapper mapper) {
+
+        return new AuctionRequestFactory(
+                maxRequestSize,
+                ortb2RequestFactory,
+                storedRequestProcessor,
+                implicitParametersExtractor,
+                auctionImplicitParametersInjector,
+                new InterstitialProcessor(),
+                ortbTypesResolver,
                 privacyEnforcementService,
+                timeoutResolver,
                 mapper);
     }
 
@@ -239,19 +264,23 @@ public class ServiceConfiguration {
 
     @Bean
     AmpRequestFactory ampRequestFactory(StoredRequestProcessor storedRequestProcessor,
-                                        AuctionRequestFactory auctionRequestFactory,
+                                        Ortb2RequestFactory ortb2RequestFactory,
                                         OrtbTypesResolver ortbTypesResolver,
                                         ImplicitParametersExtractor implicitParametersExtractor,
+                                        AuctionImplicitParametersInjector auctionImplicitParametersInjector,
                                         FpdResolver fpdResolver,
+                                        PrivacyEnforcementService privacyEnforcementService,
                                         TimeoutResolver timeoutResolver,
                                         JacksonMapper mapper) {
 
         return new AmpRequestFactory(
                 storedRequestProcessor,
-                auctionRequestFactory,
+                ortb2RequestFactory,
                 ortbTypesResolver,
                 implicitParametersExtractor,
+                auctionImplicitParametersInjector,
                 fpdResolver,
+                privacyEnforcementService,
                 timeoutResolver,
                 mapper);
     }
@@ -261,15 +290,19 @@ public class ServiceConfiguration {
             @Value("${auction.max-request-size}") int maxRequestSize,
             @Value("${video.stored-request-required}") boolean enforceStoredRequest,
             VideoStoredRequestProcessor storedRequestProcessor,
-            AuctionRequestFactory auctionRequestFactory,
+            Ortb2RequestFactory ortb2RequestFactory,
+            AuctionImplicitParametersInjector auctionImplicitParametersInjector,
+            PrivacyEnforcementService privacyEnforcementService,
             TimeoutResolver timeoutResolver,
             JacksonMapper mapper) {
 
         return new VideoRequestFactory(
                 maxRequestSize,
                 enforceStoredRequest,
+                ortb2RequestFactory,
+                auctionImplicitParametersInjector,
                 storedRequestProcessor,
-                auctionRequestFactory,
+                privacyEnforcementService,
                 timeoutResolver,
                 mapper);
     }
