+++ src/test/java/org/prebid/server/auction/requestfactory/VideoRequestFactoryTest.java
@@ -1,4 +1,4 @@
-package org.prebid.server.auction;
+package org.prebid.server.auction.requestfactory;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.iab.openrtb.request.BidRequest;
@@ -11,7 +11,6 @@ import com.iab.openrtb.request.Video;
 import com.iab.openrtb.request.video.BidRequestVideo;
 import com.iab.openrtb.request.video.PodError;
 import io.vertx.core.Future;
-import io.vertx.core.buffer.Buffer;
 import io.vertx.core.http.HttpServerRequest;
 import io.vertx.ext.web.RoutingContext;
 import org.junit.Before;
@@ -20,11 +19,20 @@ import org.junit.Test;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnit;
 import org.mockito.junit.MockitoRule;
+import org.mockito.stubbing.Answer;
 import org.prebid.server.VertxTest;
+import org.prebid.server.auction.PriceGranularity;
+import org.prebid.server.auction.PrivacyEnforcementService;
+import org.prebid.server.auction.TimeoutResolver;
+import org.prebid.server.auction.VideoStoredRequestProcessor;
 import org.prebid.server.auction.model.AuctionContext;
 import org.prebid.server.auction.model.WithPodErrors;
 import org.prebid.server.exception.InvalidRequestException;
 import org.prebid.server.metric.MetricName;
+import org.prebid.server.privacy.ccpa.Ccpa;
+import org.prebid.server.privacy.gdpr.model.TcfContext;
+import org.prebid.server.privacy.model.Privacy;
+import org.prebid.server.privacy.model.PrivacyContext;
 import org.prebid.server.proto.openrtb.ext.ExtIncludeBrandCategory;
 import org.prebid.server.proto.openrtb.ext.request.ExtRequest;
 import org.prebid.server.proto.openrtb.ext.request.ExtRequestPrebid;
@@ -35,12 +43,13 @@ import org.prebid.server.util.HttpUtil;
 
 import java.util.ArrayList;
 import java.util.Arrays;
+import java.util.List;
 
 import static java.util.Collections.emptySet;
 import static java.util.Collections.singletonList;
+import static org.apache.commons.lang3.StringUtils.EMPTY;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyList;
 import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
@@ -53,12 +62,15 @@ public class VideoRequestFactoryTest extends VertxTest {
     public final MockitoRule mockitoRule = MockitoJUnit.rule();
 
     @Mock
-    private VideoStoredRequestProcessor videoStoredRequestProcessor;
-
+    private Ortb2RequestFactory ortb2RequestFactory;
     @Mock
-    private AuctionRequestFactory auctionRequestFactory;
+    private AuctionImplicitParametersInjector implicitParametersInjector;
+    @Mock
+    private VideoStoredRequestProcessor videoStoredRequestProcessor;
+    @Mock
+    private PrivacyEnforcementService privacyEnforcementService;
 
-    private VideoRequestFactory factory;
+    private VideoRequestFactory target;
 
     @Mock
     private RoutingContext routingContext;
@@ -72,11 +84,19 @@ public class VideoRequestFactoryTest extends VertxTest {
         given(routingContext.request()).willReturn(httpServerRequest);
         given(httpServerRequest.getParam(anyString())).willReturn("test");
 
-        factory = new VideoRequestFactory(
+        final PrivacyContext defaultPrivacyContext = PrivacyContext.of(
+                Privacy.of("0", EMPTY, Ccpa.EMPTY, 0),
+                TcfContext.empty());
+        given(privacyEnforcementService.contextFromBidRequest(any()))
+                .willReturn(Future.succeededFuture(defaultPrivacyContext));
+
+        target = new VideoRequestFactory(
                 Integer.MAX_VALUE,
                 false,
+                ortb2RequestFactory,
+                implicitParametersInjector,
                 videoStoredRequestProcessor,
-                auctionRequestFactory,
+                privacyEnforcementService,
                 timeoutResolver,
                 jacksonMapper);
     }
@@ -87,7 +107,7 @@ public class VideoRequestFactoryTest extends VertxTest {
         given(routingContext.getBody()).willReturn(null);
 
         // when
-        final Future<?> future = factory.fromRequest(routingContext, 0L);
+        final Future<?> future = target.fromRequest(routingContext, 0L);
 
         // then
         assertThat(future.failed()).isTrue();
@@ -99,19 +119,21 @@ public class VideoRequestFactoryTest extends VertxTest {
     @Test
     public void shouldReturnFailedFutureIfStoredRequestIsEnforcedAndIdIsNotProvided() throws JsonProcessingException {
         // given
-        given(routingContext.getBody())
-                .willReturn(Buffer.buffer(mapper.writeValueAsBytes(BidRequestVideo.builder().build())));
+        given(routingContext.getBodyAsString())
+                .willReturn(mapper.writeValueAsString(BidRequestVideo.builder().build()));
         given(routingContext.request().getHeader(HttpUtil.USER_AGENT_HEADER)).willReturn("123");
-        factory = new VideoRequestFactory(
+        target = new VideoRequestFactory(
                 Integer.MAX_VALUE,
                 true,
+                ortb2RequestFactory,
+                implicitParametersInjector,
                 videoStoredRequestProcessor,
-                auctionRequestFactory,
+                privacyEnforcementService,
                 timeoutResolver,
                 jacksonMapper);
 
         // when
-        final Future<?> future = factory.fromRequest(routingContext, 0L);
+        final Future<?> future = target.fromRequest(routingContext, 0L);
 
         // then
         assertThat(future.failed()).isTrue();
@@ -123,18 +145,20 @@ public class VideoRequestFactoryTest extends VertxTest {
     @Test
     public void shouldReturnFailedFutureIfRequestBodyExceedsMaxRequestSize() {
         // given
-        factory = new VideoRequestFactory(
+        target = new VideoRequestFactory(
                 2,
                 true,
+                ortb2RequestFactory,
+                implicitParametersInjector,
                 videoStoredRequestProcessor,
-                auctionRequestFactory,
+                privacyEnforcementService,
                 timeoutResolver,
                 jacksonMapper);
 
-        given(routingContext.getBody()).willReturn(Buffer.buffer("body"));
+        given(routingContext.getBodyAsString()).willReturn("body");
 
         // when
-        final Future<?> future = factory.fromRequest(routingContext, 0L);
+        final Future<?> future = target.fromRequest(routingContext, 0L);
 
         // then
         assertThat(future.failed()).isTrue();
@@ -146,10 +170,10 @@ public class VideoRequestFactoryTest extends VertxTest {
     @Test
     public void shouldReturnFailedFutureIfRequestBodyCouldNotBeParsed() {
         // given
-        given(routingContext.getBody()).willReturn(Buffer.buffer("body"));
+        given(routingContext.getBodyAsString()).willReturn("body");
 
         // when
-        final Future<?> future = factory.fromRequest(routingContext, 0L);
+        final Future<?> future = target.fromRequest(routingContext, 0L);
 
         // then
         assertThat(future.failed()).isTrue();
@@ -201,39 +225,34 @@ public class VideoRequestFactoryTest extends VertxTest {
                 .ext(ExtRequest.of(ext))
                 .build();
 
-        final WithPodErrors<BidRequest> mergedBidRequest = WithPodErrors.of(
-                bidRequest, singletonList(PodError.of(1, 1, singletonList("TEST"))));
-
         final BidRequestVideo requestVideo = BidRequestVideo.builder().device(
                 Device.builder().ua("123").build()).build();
-        given(routingContext.getBody()).willReturn(Buffer.buffer(mapper.writeValueAsBytes(requestVideo)));
-        given(videoStoredRequestProcessor.processVideoRequest(any(), any(), any(), any()))
-                .willReturn(Future.succeededFuture(mergedBidRequest));
-        given(auctionRequestFactory.validateRequest(any())).willAnswer(invocation -> invocation.getArgument(0));
-        given(auctionRequestFactory.fillImplicitParameters(any(), any(), any()))
-                .willAnswer(invocation -> invocation.getArgument(0));
-        given(auctionRequestFactory.toAuctionContext(any(), any(), any(), anyList(), anyLong(), any()))
-                .willReturn(Future.succeededFuture());
+        given(routingContext.getBodyAsString()).willReturn(mapper.writeValueAsString(requestVideo));
+
+        final List<PodError> podErrors = singletonList(PodError.of(1, 1, singletonList("TEST")));
+        givenBidRequest(bidRequest, podErrors);
 
         // when
-        final Future<WithPodErrors<AuctionContext>> result = factory.fromRequest(routingContext, 0L);
+        final Future<WithPodErrors<AuctionContext>> result = target.fromRequest(routingContext, 0L);
 
         // then
-        verify(routingContext).getBody();
+        verify(routingContext).getBodyAsString();
         verify(videoStoredRequestProcessor).processVideoRequest("", null, emptySet(), requestVideo);
-        verify(auctionRequestFactory).validateRequest(bidRequest);
-        verify(auctionRequestFactory).fillImplicitParameters(bidRequest, routingContext, timeoutResolver);
-        verify(auctionRequestFactory).toAuctionContext(
-                routingContext, bidRequest, MetricName.video, new ArrayList<>(), 0, timeoutResolver);
+        verify(ortb2RequestFactory).fetchAccountAndCreateAuctionContext(
+                routingContext, bidRequest, MetricName.video, 0, new ArrayList<>());
+        verify(ortb2RequestFactory).validateRequest(bidRequest);
+        verify(implicitParametersInjector).fillImplicitParameters(bidRequest, routingContext, timeoutResolver);
+        verify(ortb2RequestFactory).enrichBidRequestWithAccountAndPrivacyData(eq(bidRequest), any(), any());
 
-        assertThat(result.result().getPodErrors()).isEqualTo(mergedBidRequest.getPodErrors());
+        assertThat(result.result().getData().getBidRequest()).isEqualTo(bidRequest);
+        assertThat(result.result().getPodErrors()).isEqualTo(podErrors);
     }
 
     @Test
     public void shouldReplaceDeviceUaWithUserAgentHeaderIfPresented() throws JsonProcessingException {
         // given
         final BidRequestVideo requestVideo = BidRequestVideo.builder().build();
-        given(routingContext.getBody()).willReturn(Buffer.buffer(mapper.writeValueAsBytes(requestVideo)));
+        given(routingContext.getBodyAsString()).willReturn(mapper.writeValueAsString(requestVideo));
         given(routingContext.request().getHeader(HttpUtil.USER_AGENT_HEADER)).willReturn("user-agent-123");
 
         final WithPodErrors<BidRequest> emptyMergeObject = WithPodErrors.of(null, null);
@@ -241,7 +260,7 @@ public class VideoRequestFactoryTest extends VertxTest {
                 .willReturn(Future.succeededFuture(emptyMergeObject));
 
         // when
-        factory.fromRequest(routingContext, 0L);
+        target.fromRequest(routingContext, 0L);
 
         // then
         verify(videoStoredRequestProcessor).processVideoRequest(any(), any(), any(), eq(BidRequestVideo.builder()
@@ -255,10 +274,10 @@ public class VideoRequestFactoryTest extends VertxTest {
     public void shouldReturnErrorIfDeviceUaAndUserAgentHeaderIsEmpty() throws JsonProcessingException {
         // given
         final BidRequestVideo requestVideo = BidRequestVideo.builder().build();
-        given(routingContext.getBody()).willReturn(Buffer.buffer(mapper.writeValueAsBytes(requestVideo)));
+        given(routingContext.getBodyAsString()).willReturn(mapper.writeValueAsString(requestVideo));
 
         // when
-        Future<WithPodErrors<AuctionContext>> future = factory.fromRequest(routingContext, 0L);
+        Future<WithPodErrors<AuctionContext>> future = target.fromRequest(routingContext, 0L);
 
         // then
         assertThat(future.failed()).isTrue();
@@ -266,4 +285,26 @@ public class VideoRequestFactoryTest extends VertxTest {
                 .isInstanceOf(InvalidRequestException.class)
                 .hasMessage("Device.UA and User-Agent Header is not presented");
     }
+
+    private void givenBidRequest(BidRequest bidRequest, List<PodError> podErrors) {
+        given(videoStoredRequestProcessor.processVideoRequest(any(), any(), any(), any()))
+                .willReturn(Future.succeededFuture(WithPodErrors.of(bidRequest, podErrors)));
+        given(ortb2RequestFactory.fetchAccountAndCreateAuctionContext(any(), any(), any(), anyLong(), any()))
+                .willAnswer(invocationOnMock -> Future.succeededFuture(
+                        AuctionContext.builder()
+                                .bidRequest((BidRequest) invocationOnMock.getArguments()[1])
+                                .build()));
+
+        given(ortb2RequestFactory.validateRequest(any())).willAnswer(answerWithFirstArgument());
+        given(implicitParametersInjector.fillImplicitParameters(any(), any(), any()))
+                .willAnswer(answerWithFirstArgument());
+
+        given(ortb2RequestFactory.enrichBidRequestWithAccountAndPrivacyData(any(), any(), any()))
+                .willAnswer(answerWithFirstArgument());
+    }
+
+    private Answer<Object> answerWithFirstArgument() {
+        return invocationOnMock -> invocationOnMock.getArguments()[0];
+    }
+
 }
